# refactor: Use strings.Builder