When used for appending strings, strings.Builder yields fewer
allocations and is generally faster than bytes.Buffer.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkQuery_String-4             3219          2899          -9.94%
BenchmarkExprNames-4                6655          6646          -0.14%
BenchmarkEval-4                     519           522           +0.58%
BenchmarkParserParseStatement-4     6768          6083          -10.12%
BenchmarkSanitize-4                 5846          5525          -5.49%

benchmark                           old MB/s     new MB/s     speedup
BenchmarkParserParseStatement-4     6.65         7.40         1.11x

benchmark                           old allocs     new allocs     delta
BenchmarkQuery_String-4             44             39             -11.36%
BenchmarkExprNames-4                10             10             +0.00%
BenchmarkEval-4                     6              6              +0.00%
BenchmarkParserParseStatement-4     56             44             -21.43%
BenchmarkSanitize-4                 10             7              -30.00%

benchmark                           old bytes     new bytes     delta
BenchmarkQuery_String-4             1336          736           -44.91%
BenchmarkExprNames-4                240           240           +0.00%
BenchmarkEval-4                     80            80            +0.00%
BenchmarkParserParseStatement-4     6448          5320          -17.49%
BenchmarkSanitize-4                 1204          769           -36.13%
```