@dgnorton 2020-08-28
@ayang64 looks like a good change. The tests are failing because the build is using Go 1.9.2. Can you create a separate PR to bump the version of Go up and then rebase this once that merges?